+++ client/index.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 // import App from './components/app';
-import App2 from './components/App2';
+import App from './components/App';
 
-// We modified the render to go to App2 , like a sandbox to see if we can figure it out 
+// We modified the render to go to App , like a sandbox to see if we can figure it out 
 ReactDOM.render(
-    <App2 />, 
+    <App />, 
     document.getElementById('root')
 );
